html/rendering/non-replaced-elements/the-fieldset-and-legend-elements/fieldset-multicol.html

<!doctype html>
<title>fieldset multicol</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<style>
 #test { margin: 0; padding: 0; border: none }
 #test, #ref, #appearance-test { columns: 5 }
 p { margin: 0 }
 #appearance-test { -webkit-appearance: fieldset; }
</style>
<fieldset id=test>
  <p>1
  <p>2
  <p>3
  <p>4
  <p>5
</fieldset>
<div id=ref>
  <p>1
  <p>2
  <p>3
  <p>4
  <p>5
</div>
<div id=appearance-test>
  <p>1
  <p>2
  <p>3
  <p>4
  <p>5
</div>
<script>
  test(() => {
    assert_equals(getComputedStyle(document.getElementById('test')).height,
                  getComputedStyle(document.getElementById('ref')).height);
  });

  test(() => {
    assert_equals(getComputedStyle(document.getElementById('appearance-test')).height,
                  getComputedStyle(document.getElementById('ref')).height);
  }, '-webkit-appearance: fieldset and multicol');
</script>
