third_party/blink/web_tests/flag-specific/disable-layout-ng/external/wpt/html/rendering/non-replaced-elements/the-fieldset-and-legend-elements/fieldset-multicol-expected.txt

This is a testharness.js-based test.
FAIL fieldset multicol assert_equals: expected "20px" but got "100px"
Harness: the test ran to completion.
